Fix baudrate variable | Fix Comm Reset | Fix Timer Issue

+++ CANopenNode_STM32/CO_app_STM32.h
@@ -31,7 +31,7 @@ typedef struct {
 	 * be the final NodeID, after calling canopen_app_init() you should check ActiveNodeID of CANopenNodeSTM32 structure for assigned Node ID.
 	 */
     uint8_t activeNodeID; /* Assigned Node ID */
-    uint8_t baudrate;     /* This is the baudrate you've set in your CubeMX Configuration */
+    uint16_t baudrate;     /* This is the baudrate you've set in your CubeMX Configuration */
     TIM_HandleTypeDef*
         timerHandle; /*Pass in the timer that is going to be used for generating 1ms interrupt for tmrThread function,
 	 * please note that CANOpenSTM32 Library will override HAL_TIM_PeriodElapsedCallback function, if you also need this function in your codes, please take required steps
